Add: カスタム Hooks にまとめ直し
