+++ src/components/organisms/IOSPermissionsCheckModal/index.ts
@@ -1,1 +1,2 @@
 export { IOSPermissionsCheckModal } from "./IOSPermissionsCheckModal";
+export { useIOSPermissionsCheckModal } from "./IOSPermissionsCheckModal.hooks";
